Completed migration of Queue and Stack
