--- src/D/QueueAndStack/README.md
@@ -1,5 +1,0 @@
-## Queue & Stack
-1.  [Sort with 3 Stacks](Medium/SortWith3Stacks)
-2.  [Queue by Two Stacks](Medium/QueueByTwoStacks)
-3.  [Stack with min()](Medium/StackWithMin)
-4.  [Sort with 2 Stacks](Medium/SortWith2Stacks)
